requirements.txt
Pillow
# FakeRPi # https://github.com/sn4k3/FakeRPi
epd-library
schedule